build: update Gradle version and optimize Java setup
Updated the Gradle version from 8.1.1 to 8.2.1 in gradle.properties and gradle-wrapper.properties. This upgrade is important to accommodate latest features and improvements in Gradle. Modified gradlew file to check for 'java' command availability in the system path. This ensures that an error message is displayed when JAVA_HOME is not set and 'java' command is not available in PATH. Furthermore, validated the distribution URL in gradle-wrapper.properties, which adds an additional layer of security during Gradle Wrapper setup.