build: remove Dokka plugin from Gradle configuration
Dokka plugin was removed from the Gradle configuration file. This is part of the ongoing effort to simplify our build process by removing unused or unnecessary dependencies. It was determined that this library is no longer needed in our project. This change will help to speed up our builds and reduce project complexity.